requirements.txt
pip==24.1.2
fastapi==0.115.8
fastapi-cli==0.0.7
alembic==1.15.1
psycopg==3.2.6
psycopg2==2.9.10
python-jose==3.5.0 
sqlmodel==0.0.24
SQLAlchemy==2.0.39
rich==13.9.2
PyYAML==6.0.2
passlib==1.7.4
cryptography==44.0.0
uvicorn==0.34.0